Ensure that purge and delete get deleted when blue text protection is on

+++ tg_bot/modules/msg_deleting.py
@@ -1,5 +1,5 @@
 from telegram.error import BadRequest
-from telegram.ext import CommandHandler, Filters
+from telegram.ext import CommandHandler
 from telegram.ext.dispatcher import run_async
 
 from tg_bot import dispatcher, LOGGER
@@ -9,27 +9,33 @@ from tg_bot.modules.helper_funcs.chat_status import user_admin, can_delete
 @run_async
 @user_admin
 def purge(bot, update):
-    curr_message = update.effective_message
-    chat = update.effective_chat
-    if can_delete(chat, bot.id):
-        message_id = curr_message.reply_to_message.message_id
-        curr_message_id = curr_message.message_id
-        for m_id in range(message_id, curr_message_id + 1):  # +1 to include curr message
-            try:
-                bot.deleteMessage(chat.id, m_id)
-            except BadRequest as err:
-                if err.message != "Message to delete not found":
-                    LOGGER.exception("Error while purging chat messages.")
-        bot.send_message(chat.id, "Purge complete.", 'Markdown')
+    msg = update.effective_message
+    if msg.reply_to_message:
+        chat = update.effective_chat
+        if can_delete(chat, bot.id):
+            message_id = msg.reply_to_message.message_id
+            curr_message_id = msg.message_id
+            for m_id in range(message_id, curr_message_id + 1):  # +1 to include curr message
+                try:
+                    bot.deleteMessage(chat.id, m_id)
+                except BadRequest as err:
+                    if err.message != "Message to delete not found":
+                        LOGGER.exception("Error while purging chat messages.")
+            bot.send_message(chat.id, "Purge complete.")
+    else:
+        msg.reply_to_message("Reply to a message to select where to start purging from.")
 
 
 @run_async
 @user_admin
 def del_message(bot, update):
-    chat = update.effective_chat
-    if can_delete(chat, bot.id):
-        update.effective_message.reply_to_message.delete()
-        update.effective_message.delete()
+    if update.effective_message.reply_to_message:
+        chat = update.effective_chat
+        if can_delete(chat, bot.id):
+            update.effective_message.reply_to_message.delete()
+            update.effective_message.delete()
+    else:
+        update.effective_message.reply_to_message("Whadya want to delete?")
 
 
 __help__ = """
@@ -41,8 +47,8 @@ __help__ = """
 __name__ = "Purges"
 
 
-DELETE_HANDLER = CommandHandler("del", del_message, filters=Filters.reply)
-PURGE_HANDLER = CommandHandler("purge", purge, filters=Filters.reply)
+DELETE_HANDLER = CommandHandler("del", del_message)
+PURGE_HANDLER = CommandHandler("purge", purge)
 
 dispatcher.add_handler(DELETE_HANDLER)
 dispatcher.add_handler(PURGE_HANDLER)
